+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- pdf_pcb_print.output can be used instead of pdf_pcb_print.output_name
 
 ## [0.5.0] - 2020-07-11
 ### Changed